add insert command with modes
